fix warnings and empty tests
This is purely before any work on latency adjusting, and is just about
geting the tests to run without warnings or skipped tests. Getting this
involved:
1. renaming some lingering examples of p to probs and `q` to
`quantile_levels` (and some `quantile_level`s to the plural)
2. Adding snapshots so the tests for printing in population_scaling work
as intended (should probably be converted to cli_informs at some point)
3. removing the nearly empty `test-propagate_samples` which seems like
something intended that was never finished. Probably want to add an
issue if we actually want it done.
4. added a bunch of `edf`'s in unhappy `prep` steps